Added extended mode, allowing requests to do more.
Extended mode allows several containers to be updated with a
single request to the server. This is possible by having the
server do a bit more work and return a more meaningful response.

To go into extended mode don't specify a container when making
the request. In this case, the server needs to return a JSON response
specifying which containers to update, what page title to set, etc.

Normal mode should work as it used to, without being affected by this.