Added beforeSuccess option to $.pjax.
The callback will be executed before $.pjax applies the response data
for a given request, but only if the response data looks valid.
This allows you to do some stuff before the content in the container
gets replaced.
If the beforeSuccess callback returns false, pjax will not perform
any changes to the page.