Added a 'fragment' option to support servers that always return full HTML pages
The 'fragment' option allows specifying a string selector which selects an element to be pulled out of the received response HTML. This is useful if you have no influence on the server side or do not want to modify the backend application. For example, 'fragment: "#content"' pulls the content element out of the response and inserts its children into the container element that you chose.