Move pjax:receive event right before content change
As mentioned in a comment [1], we want this event to be fired closer to
the actual DOM change. A subsequent change might want to make this event
cancelable, so that you can provide your own content replacement
technique. But for now we're keeping it simple and only allowing
listeners to react to the event without interfering.

[1]: https://github.com/defunkt/jquery-pjax/pull/384#issuecomment-42300835