Rename pjax:receive => pjax:beforeReplace
As pointed out by @josh in a comment to PR #384 [1], it would make more
sense if this event is called something other than "pjax:receive". We
initially picked that name to stay consistent with Turbolinks event
naming, but now that we've moved the event to a slightly different place
in the pjax lifecycle, the name could be more descriptive. Looking at
the names of the events in the readme, we think that
"pjax:beforeReplace" is a good descriptive name.

[1]: https://github.com/defunkt/jquery-pjax/pull/384#issuecomment-42305547