Send the content to pjax:beforeReplace
Before, we would send different data along with the pjax:beforeReplace
depending on when the event was fired. When fired in the context of an
AJAX request, the response `data` from the server would be passed along.
In the context of a popstate event, we didn't have access to the
response from the server, so we just passed along `null` in place of the
server response data.

@josh and @mislav suggested that we pass along the contents that we are
about to inject into the pjax container. We think that's a good idea,
it will allow people to inspect the new content before it's replaced.
And when we make this event cancelable in the future, we could allow
people to implement their own content-replacement strategy.