Fix docs stating that `body` is the default value for `fragment`
That's not exactly true, as we require users to specify an explicit
`fragment` value if they're dealing with full-HTML responses from the
server. This explicit value can be "body" if the user chooses so, but
that must be decided on a per-app basis.

If no explicit `fragment` value is provided, a full-HTML response will
trigger a native reload of the page.

Closes #475 [ci skip]