Decode hash value before querying for named anchor
With a URL like `#%24.ajax`, we need to first decode the value before
searching the DOM for the element named `$.ajax`.

In Firefox, this is not necessary, as the browser already auto-decodes
hash values. Therefore, in Firefox this will perform double-decoding,
which could theoretically lead to inability to match anchors that
contain a percent-encoded value literally. However, since Firefox also
seems to automatically decode values for anchors named via `name="..."`
HTML attribute, you're going to have a bad time anyway with named
anchors that contain percent-encoded characters literally, with or
without pjax.