Keep hash in temporarily changed URL while the page is loading
If you navigate with pjax to `/foo.html#bar`, the URL bar will change to
`/foo.html` while the page is loading, but only add `#bar` when it's
finished. There's some debate about whether we should be changing the
URL at all while loading, because browsers don't seem to do it until the
page starts sending data back. However, while we're still changing the
URL, it's worth ensuring that at least that URL is complete.