Fix executing SCRIPT tags on pjax back/forward
After jQuery eval's a SCRIPT tag for the first time, it tags it with an
internal flag "globalEval" that is kept even after cloning and prevents
the same script from getting executed again.

Now reset that property when writing to pjax cache, so that back/forward
navigation properly re-executes inline SCRIPT tags.