Fixed scrollTo behavior.
It was not previously considering the cases when the immediate
vertical scrolling is not a applied to the window, but to a scrollable
parent element.

The scrolling to a hash target has to also be performed separately,
because the hash target may have a different scrollable parent from the
container contents.

For cross-browser comptability, there is a distinction in the code
between regular elements and "root" elements ( i.e. `window`,
`document`, `html`, & `body` ). These elements require
`$(window).scrollTop()` as a getter and `$('body, html').scrollTop(n)`
as a setter for the `scrollTop` value.